Refactor into core and json module, add orjson support
