--- src/pythonjsonlogger/jsonlogger.py
@@ -1,304 +1,0 @@
-"""
-This library is provided to allow standard python logging
-to output log data as JSON formatted strings
-"""
-
-import logging
-import json
-import re
-import traceback
-import importlib
-from datetime import date, datetime, time, timezone
-import sys
-from typing import Any, Callable, Dict, List, Optional, Tuple, Union
-
-from inspect import istraceback
-
-from collections import OrderedDict
-
-# skip natural LogRecord attributes
-# http://docs.python.org/library/logging.html#logrecord-attributes
-# Changed in 3.0.0, is now list[str] instead of tuple[str, ...]
-RESERVED_ATTRS: List[str] = [
-    "args",
-    "asctime",
-    "created",
-    "exc_info",
-    "exc_text",
-    "filename",
-    "funcName",
-    "levelname",
-    "levelno",
-    "lineno",
-    "module",
-    "msecs",
-    "message",
-    "msg",
-    "name",
-    "pathname",
-    "process",
-    "processName",
-    "relativeCreated",
-    "stack_info",
-    "thread",
-    "threadName",
-]
-
-if sys.version_info >= (3, 12):
-    # taskName added in python 3.12
-    RESERVED_ATTRS.append("taskName")
-    RESERVED_ATTRS.sort()
-
-OptionalCallableOrStr = Optional[Union[Callable, str]]
-
-
-def merge_record_extra(
-    record: logging.LogRecord,
-    target: Dict,
-    reserved: Union[Dict, List],
-    rename_fields: Optional[Dict[str, str]] = None,
-) -> Dict:
-    """
-    Merges extra attributes from LogRecord object into target dictionary
-
-    :param record: logging.LogRecord
-    :param target: dict to update
-    :param reserved: dict or list with reserved keys to skip
-    :param rename_fields: an optional dict, used to rename field names in the output.
-            Rename levelname to log.level: {'levelname': 'log.level'}
-    """
-    if rename_fields is None:
-        rename_fields = {}
-    for key, value in record.__dict__.items():
-        # this allows to have numeric keys
-        if key not in reserved and not (hasattr(key, "startswith") and key.startswith("_")):
-            target[rename_fields.get(key, key)] = value
-    return target
-
-
-class JsonEncoder(json.JSONEncoder):
-    """
-    A custom encoder extending the default JSONEncoder
-    """
-
-    def default(self, o: Any) -> Any:
-        if isinstance(o, (date, datetime, time)):
-            return self.format_datetime_obj(o)
-
-        if istraceback(o):
-            return "".join(traceback.format_tb(o)).strip()
-
-        # pylint: disable=unidiomatic-typecheck
-        if type(o) == Exception or isinstance(o, Exception) or type(o) == type:
-            return str(o)
-
-        try:
-            return super().default(o)
-
-        except TypeError:
-            try:
-                return str(o)
-
-            except Exception:  # pylint: disable=broad-exception-caught
-                return None
-
-    def format_datetime_obj(self, o):
-        """Format datetime objects found in self.default
-
-        This allows subclasses to change the datetime format without understanding the
-        internals of the default method.
-        """
-        return o.isoformat()
-
-
-class JsonFormatter(logging.Formatter):
-    """
-    A custom formatter to format logging records as json strings.
-    Extra values will be formatted as str() if not supported by
-    json default encoder
-    """
-
-    # pylint: disable=too-many-arguments
-    def __init__(
-        self,
-        *args: Any,
-        json_default: OptionalCallableOrStr = None,
-        json_encoder: OptionalCallableOrStr = None,
-        json_serializer: Union[Callable, str] = json.dumps,
-        json_indent: Optional[Union[int, str]] = None,
-        json_ensure_ascii: bool = True,
-        prefix: str = "",
-        rename_fields: Optional[dict] = None,
-        static_fields: Optional[dict] = None,
-        reserved_attrs: Union[Tuple[str, ...], List[str], None] = None,
-        timestamp: Union[bool, str] = False,
-        **kwargs: Any,
-    ) -> None:
-        """
-        :param json_default: a function for encoding non-standard objects
-            as outlined in https://docs.python.org/3/library/json.html
-        :param json_encoder: optional custom encoder
-        :param json_serializer: a :meth:`json.dumps`-compatible callable
-            that will be used to serialize the log record.
-        :param json_indent: indent parameter for json.dumps
-        :param json_ensure_ascii: ensure_ascii parameter for json.dumps
-        :param prefix: an optional string prefix added at the beginning of
-            the formatted string
-        :param rename_fields: an optional dict, used to rename field names in the output.
-            Rename message to @message: {'message': '@message'}
-        :param static_fields: an optional dict, used to add fields with static values to all logs
-        :param reserved_attrs: an optional list of fields that will be skipped when
-            outputting json log record. Defaults to all log record attributes:
-            http://docs.python.org/library/logging.html#logrecord-attributes
-        :param timestamp: an optional string/boolean field to add a timestamp when
-            outputting the json log record. If string is passed, timestamp will be added
-            to log record using string as key. If True boolean is passed, timestamp key
-            will be "timestamp". Defaults to False/off.
-        """
-        self.json_default = self._str_to_fn(json_default)
-        self.json_encoder = self._str_to_fn(json_encoder)
-        self.json_serializer = self._str_to_fn(json_serializer)
-        self.json_indent = json_indent
-        self.json_ensure_ascii = json_ensure_ascii
-        self.prefix = prefix
-        self.rename_fields = rename_fields or {}
-        self.static_fields = static_fields or {}
-        if reserved_attrs is None:
-            reserved_attrs = RESERVED_ATTRS
-        self.reserved_attrs = dict(zip(reserved_attrs, reserved_attrs))
-        self.timestamp = timestamp
-
-        # super(JsonFormatter, self).__init__(*args, **kwargs)
-        logging.Formatter.__init__(self, *args, **kwargs)
-        if not self.json_encoder and not self.json_default:
-            self.json_encoder = JsonEncoder
-
-        self._required_fields = self.parse()
-        self._skip_fields = dict(zip(self._required_fields, self._required_fields))
-        self._skip_fields.update(self.reserved_attrs)
-        return
-
-    def _str_to_fn(self, fn_as_str):
-        """
-        If the argument is not a string, return whatever was passed in.
-        Parses a string such as package.module.function, imports the module
-        and returns the function.
-
-        :param fn_as_str: The string to parse. If not a string, return it.
-        """
-        if not isinstance(fn_as_str, str):
-            return fn_as_str
-
-        path, _, function = fn_as_str.rpartition(".")
-        module = importlib.import_module(path)
-        return getattr(module, function)
-
-    def parse(self) -> List[str]:
-        """
-        Parses format string looking for substitutions
-
-        This method is responsible for returning a list of fields (as strings)
-        to include in all log messages.
-        """
-        if isinstance(self._style, logging.StringTemplateStyle):
-            formatter_style_pattern = re.compile(r"\$\{(.+?)\}", re.IGNORECASE)
-        elif isinstance(self._style, logging.StrFormatStyle):
-            formatter_style_pattern = re.compile(r"\{(.+?)\}", re.IGNORECASE)
-        # PercentStyle is parent class of StringTemplateStyle and StrFormatStyle so
-        # it needs to be checked last.
-        elif isinstance(self._style, logging.PercentStyle):
-            formatter_style_pattern = re.compile(r"%\((.+?)\)", re.IGNORECASE)
-        else:
-            raise ValueError(f"Invalid format: {self._fmt!r}")
-
-        if self._fmt:
-            return formatter_style_pattern.findall(self._fmt)
-        return []
-
-    def add_fields(
-        self,
-        log_record: Dict[str, Any],
-        record: logging.LogRecord,
-        message_dict: Dict[str, Any],
-    ) -> None:
-        """
-        Override this method to implement custom logic for adding fields.
-        """
-        for field in self._required_fields:
-            log_record[field] = record.__dict__.get(field)
-
-        log_record.update(self.static_fields)
-        log_record.update(message_dict)
-        merge_record_extra(
-            record,
-            log_record,
-            reserved=self._skip_fields,
-            rename_fields=self.rename_fields,
-        )
-
-        if self.timestamp:
-            # TODO: Can this use isinstance instead?
-            # pylint: disable=unidiomatic-typecheck
-            key = self.timestamp if type(self.timestamp) == str else "timestamp"
-            log_record[key] = datetime.fromtimestamp(record.created, tz=timezone.utc)
-
-        self._perform_rename_log_fields(log_record)
-        return
-
-    def _perform_rename_log_fields(self, log_record: Dict[str, Any]) -> None:
-        for old_field_name, new_field_name in self.rename_fields.items():
-            log_record[new_field_name] = log_record[old_field_name]
-            del log_record[old_field_name]
-        return
-
-    def process_log_record(self, log_record: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Override this method to implement custom logic
-        on the possibly ordered dictionary.
-        """
-        return log_record
-
-    def jsonify_log_record(self, log_record: Dict[str, Any]) -> str:
-        """Returns a json string of the log record."""
-        return self.json_serializer(
-            log_record,
-            default=self.json_default,
-            cls=self.json_encoder,
-            indent=self.json_indent,
-            ensure_ascii=self.json_ensure_ascii,
-        )
-
-    def serialize_log_record(self, log_record: Dict[str, Any]) -> str:
-        """Returns the final representation of the log record."""
-        return self.prefix + self.jsonify_log_record(log_record)
-
-    def format(self, record: logging.LogRecord) -> str:
-        """Formats a log record and serializes to json"""
-        message_dict: Dict[str, Any] = {}
-        # TODO: logging.LogRecord.msg and logging.LogRecord.message in typeshed
-        #        are always type of str. We shouldn't need to override that.
-        if isinstance(record.msg, dict):
-            message_dict = record.msg
-            record.message = ""
-        else:
-            record.message = record.getMessage()
-        # only format time if needed
-        if "asctime" in self._required_fields:
-            record.asctime = self.formatTime(record, self.datefmt)
-
-        # Display formatted exception, but allow overriding it in the
-        # user-supplied dict.
-        if record.exc_info and not message_dict.get("exc_info"):
-            message_dict["exc_info"] = self.formatException(record.exc_info)
-        if not message_dict.get("exc_info") and record.exc_text:
-            message_dict["exc_info"] = record.exc_text
-        # Display formatted record of stack frames
-        # default format is a string returned from :func:`traceback.print_stack`
-        if record.stack_info and not message_dict.get("stack_info"):
-            message_dict["stack_info"] = self.formatStack(record.stack_info)
-
-        log_record: Dict[str, Any] = OrderedDict()
-        self.add_fields(log_record, record, message_dict)
-        log_record = self.process_log_record(log_record)
-
-        return self.serialize_log_record(log_record)
