+++ tox.ini
@@ -17,19 +17,19 @@ python =
 
 [testenv]
 description = run unit tests
-extras = test
+extras = dev
 commands =
     pytest tests
 
 [testenv:format]
 description = run formatters
-extras = lint
+extras = dev
 commands =
     black src tests
 
 [testenv:lint]
 description = run linters
-extras = lint
+extras = dev
 commands =
     validate-pyproject pyproject.toml
     black --check --diff src tests
